Updated crates, Rust and code
New release which uses the latest versions of crates and Rust.

- Use latest Rust edition 2021
- Replaced clap with pico-args (Smaller binary, less crates)
- Updated mpris crate
- Strip binary via cargo profile
- Use cargo fmt to format the code
- Added workflow to build and upload binary to releases

Note: The previous version didn't allowed the `subcommands` to be used
as the first argument, but it now can! It still supports the old format
to be backwards compatible, but i prefer to use the subcommand as the
first argument my self.